chore: refactor query page
